moving zh inventories for allowing new grouppings
